refactor: rename logging → logge to avoid shadowing
