+++ analyses/bhadra/25_07_02_pipeline/compute_frozen_features_neuroprobe.py
@@ -13,7 +13,7 @@ from torch.optim.lr_scheduler import ChainedScheduler
 from bfm.training.training_config import convert_dtypes, unconvert_dtypes, parse_subject_trials_from_config
 from torch.utils.data import DataLoader
 from bfm.training.setup_registry import setups
-from bfm.core.logging import log
+from bfm.core.logger import log
 
 from bfm.evaluation.neuroprobe.datasets import BrainTreebankSubjectTrialBenchmarkDataset
 import evaluation.neuroprobe.config as neuroprobe_config
