+++ bfm/evaluation/neuroprobe_frozen.py
@@ -13,7 +13,7 @@ import torch
 import torch.cuda
 from torch.utils.data import DataLoader
 
-from bfm.core.logging import log
+from bfm.core.logger import log
 from bfm.evaluation.neuroprobe.train_test_splits import generate_splits_SS_SM
 import evaluation.neuroprobe.config as neuroprobe_config
 from bfm.evaluation.neuroprobe.datasets import BrainTreebankSubjectTrialBenchmarkDataset
