+++ bfm/pretrain.py
@@ -19,7 +19,7 @@ from bfm.training.optimizer.builders import build_optimizers, build_schedulers
 from bfm.subject.datasets.dataset import load_subjects
 from bfm.evaluation.neuroprobe_tasks import FrozenModelEvaluation_SS_SM
 from bfm.training.setup_registry import setups
-from bfm.core.logging import log
+from bfm.core.logger import log
 from bfm.training.training_config import (
     update_dir_name,
     update_random_seed,
