+++ bfm/training/setups/mse_rm.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 
 from bfm.training.setup_registry import setups
-from bfm.core.logging import log
+from bfm.core.logger import log
 from bfm.training.training_setup import TrainingSetup
 
 from bfm.model.base import BFModule
